requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-http-client==3.3.4
sendgrid==6.9.2
six==1.16.0
starkbank-ecdsa==2.0.3
Werkzeug==2.0.2
Flask-ReCaptcha==0.4.2
waitress
python-dotenv